last ar views and last voiceover
the final ar views are in,. as a result of the coronavirus we have not been able to visit the location to set up our reality composer scene. nonetheless we have set up the AR scenes as little dioramas that the user can set up and look at on any flat surface.